requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiosqlite==0.20.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
Crawl4AI==0.3.742
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.14
dirtyjson==1.0.8
distlib==0.3.9
distro==1.9.0
docstring_parser==0.16
duckduckgo_search==6.3.2
fake-http-header==0.3.5
fastapi==0.115.2
filelock==3.16.1
fonttools==4.54.1
frozendict==2.4.6
frozenlist==1.4.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
groq==0.13.1
h11==0.14.0
html2text==2024.2.26
html5lib==1.1
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
identify==2.6.1
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
litellm==1.52.16
llama-index-core==0.11.19
llama-index-llms-ollama==0.3.4
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
matplotlib==3.9.2
mdurl==0.1.2
mockito==1.5.3
mplfinance==0.12.10b0
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
narwhals==1.18.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
nodeenv==1.9.1
numpy==1.26.4
ollama==0.3.3
openai==1.55.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
peewee==3.17.7
pillow==10.4.0
platformdirs==4.3.6
playwright==1.47.0
pluggy==1.5.0
pre_commit==4.0.1
primp==0.6.4
propcache==0.2.0
protobuf==5.29.1
psycopg2-binary==2.9.9
py4j==0.10.9.7
pyarrow==18.1.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pyee==12.0.0
Pygments==2.18.0
pyparsing==3.2.0
PySocks==1.7.1
pyspark==3.5.4
pytest==8.3.3
pytest-mockito==0.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.2
rpds-py==0.21.0
scipy==1.14.1
selenium==4.25.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.39.2
streamlit==1.41.1
swarm @ git+https://github.com/openai/swarm.git@9db581cecaacea0d46a933d6453c312b034dbf47
tf-playwright-stealth==1.1.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
tornado==6.4.2
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.5
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
virtualenv==20.27.0
watchdog==6.0.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.15.2
yfinance==0.2.44
zipp==3.21.0
